tests/requirements.txt
pytest
coverage[toml]
furo
sphinx-immaterial
-r ../requirements-vcs.txt
